Move gitignore code to gitignore.rs
